requirements.txt
django
django-crispy-forms
Pillow
python-dotenv
gunicorn
psycopg2-binary
dj-database-url
whitenoise[brotli]
